# controllers/auth/register.go
package auth

import (
	"github.com/Felix-kuang/todo-golang-api/db"
	"github.com/Felix-kuang/todo-golang-api/models"

	"github.com/gin-gonic/gin"
	"golang.org/x/crypto/bcrypt"
)

// Register godoc
// @Summary Register a new user
// @Description Register by username & password
// @Tags Auth
// @Accept  json
// @Produce  json
// @Param   input  body  models.SwaggerRegisterInput true  "User credentials"
// @Success 200 {object} models.SwaggerSuccessRegisterResponse
// @Failure 400 {object} models.SwaggerErrorResponse
// @Router /auth/register [post]
func Register(c *gin.Context) {
	var input struct {
		Username        string `json:"username"`
		Password        string `json:"password"`
		ConfirmPassword string `json:"confirmPassword"`
	}

	if err := c.ShouldBindJSON(&input); err != nil {
		c.JSON(400, gin.H{"error": "Invalid Input"})
		return
	}

	if input.Password != input.ConfirmPassword {
		c.JSON(400, gin.H{"error": "Password Confirmation doesn't match"})
		return
	}

	//Hash the password
	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(input.Password), bcrypt.DefaultCost)

	if err != nil {
		c.JSON(500, gin.H{"error": "Could not hash password"})
		return
	}

	user := models.User{
		Username: input.Username,
		Password: string(hashedPassword),
	}

	if err := db.DbClient.Create(&user).Error; err != nil {
		c.JSON(500, gin.H{"error": "Failed to create user"})
		return
	}

	c.JSON(200, gin.H{
		"message": "User Registered successfully",
		"user": gin.H{
			"id":       user.ID,
			"username": user.Username,
		},
	})
}